+++ Kernel/Output/HTML/Layout.pm
@@ -5655,11 +5655,11 @@ sub SetRichTextParameters {
     else {
         @Toolbar = [
             [
-                'Bold',          'Italic',       'Underline',      'Strike', '-',    'NumberedList',
+                'Bold',          'Italic',       'Underline', 'Table', 'Strike', '-',    'NumberedList',
                 'BulletedList',  '-',            'Outdent',        'Indent', '-',    'JustifyLeft',
                 'JustifyCenter', 'JustifyRight', 'JustifyBlock',   '-',      'Link', 'Unlink',
                 '-',             'Image',        'HorizontalRule', '-',      'Undo', 'Redo',
-                '-',             'Find',         'SpellCheck'
+                '-',             'Find',         'SpellCheck', 'PasteText'
             ],
             '/',
             [
@@ -5670,13 +5670,13 @@ sub SetRichTextParameters {
         ];
         @ToolbarWithoutImage = [
             [
-                'Bold',          'Italic',       'Underline',    'Strike',
+                'Bold',          'Italic',       'Underline', 'Table', 'Strike',
                 '-',             'NumberedList', 'BulletedList', '-',
                 'Outdent',       'Indent',       '-',            'JustifyLeft',
                 'JustifyCenter', 'JustifyRight', 'JustifyBlock', '-',
                 'Link',          'Unlink',       '-',            'HorizontalRule',
                 '-',             'Undo',         'Redo',         '-',
-                'Find',          'SpellCheck'
+                'Find',          'SpellCheck', 'PasteText'
             ],
             '/',
             [
@@ -5786,8 +5786,8 @@ sub CustomerSetRichTextParameters {
                 'Bold',          'Italic',       'Underline',      'Strike', '-',    'NumberedList',
                 'BulletedList',  '-',            'Outdent',        'Indent', '-',    'JustifyLeft',
                 'JustifyCenter', 'JustifyRight', 'JustifyBlock',   '-',      'Link', 'Unlink',
-                '-',             'Image',        'HorizontalRule', '-',      'Undo', 'Redo',
-                '-',             'Find',         'SpellCheck'
+                '-',             'Image',        'HorizontalRule', 'Table', '-',      'Undo', 'Redo',
+                '-',             'Find',         'SpellCheck', 'PasteText'
             ],
             '/',
             [
@@ -5802,9 +5802,9 @@ sub CustomerSetRichTextParameters {
                 '-',             'NumberedList', 'BulletedList', '-',
                 'Outdent',       'Indent',       '-',            'JustifyLeft',
                 'JustifyCenter', 'JustifyRight', 'JustifyBlock', '-',
-                'Link',          'Unlink',       '-',            'HorizontalRule',
+                'Link',          'Unlink',       '-',            'HorizontalRule', 'Table',
                 '-',             'Undo',         'Redo',         '-',
-                'Find',          'SpellCheck'
+                'Find',          'SpellCheck', 'PasteText'
             ],
             '/',
             [